# DeBroglieProject/docs/swagger.yaml
basePath: /api
definitions:
  ds.Particle:
    properties:
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      mass:
        type: number
      name:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      description:
        type: string
      status:
        type: string
    type: object
...
  handler.successResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для расчета длины волны де Бройля частиц
  title: De Broglie Project API
  version: "1.0"
paths:
  /debrogliecalculations/{id}/particle/{particleId}:
    delete:
      description: Удаляет частицу из черновика заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID частицы
        in: path
        name: particleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверные параметры или статус заявки
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка или частица не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удаление частицы из заявки
      tags:
      - DeBroglieCalculations
    put:
      consumes:
      - application/json
      description: Обновляет скорость частицы в расчете де Бройля
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID частицы
        in: path
        name: particleId
        required: true
        type: integer
      - description: Новая скорость
        in: body
        name: request
        required: true
        schema:
          properties:
            speed:
              type: number
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверные параметры или статус заявки
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Заявка не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление скорости частицы
      tags:
      - DeBroglieCalculations
  /particles:
    get:
      description: Возвращает список всех частиц с возможностью фильтрации по названию
      parameters:
      - description: Название частицы для фильтрации
        in: query
        name: particle
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список частиц
          schema:
            items:
              $ref: '#/definitions/ds.Particle'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Получение списка частиц
      tags:
      - Particles
    post:
      consumes:
      - application/json
      description: Создает новую частицу (только для модераторов)
      parameters:
      - description: Данные частицы
        in: body
        name: request
        required: true
        schema:
          properties:
            description:
              type: string
            mass:
              type: number
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Созданная частица
          schema:
            $ref: '#/definitions/ds.Particle'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Создание новой частицы
      tags:
      - Particles
  /particles/{id}:
    delete:
      description: Удаляет частицу и связанное с ней изображение (только для модераторов)
      parameters:
      - description: ID частицы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Частица не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Удаление частицы
      tags:
      - Particles
    get:
      description: Возвращает информацию о частице по её ID
      parameters:
      - description: ID частицы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о частице
          schema:
            $ref: '#/definitions/ds.Particle'
        "400":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Частица не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Получение частицы по ID
      tags:
      - Particles
    put:
      consumes:
      - application/json
      description: Обновляет информацию о частице (только для модераторов)
      parameters:
      - description: ID частицы
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные частицы
        in: body
        name: request
        required: true
        schema:
          properties:
            description:
              type: string
            mass:
              type: number
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Частица не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Обновление частицы
      tags:
      - Particles
  /particles/{id}/addParticle:
    post:
      description: Добавляет частицу в черновик заявки на расчет де Бройля
      parameters:
      - description: ID частицы
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное добавление
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный ID или частица уже добавлена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Добавление частицы в заявку
      tags:
      - Particles
  /particles/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для частицы в MinIO (только для модераторов)
      parameters:
      - description: ID частицы
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Успешная загрузка
          schema:
            $ref: '#/definitions/handler.successResponse'
        "400":
          description: Неверный формат запроса
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Требуется авторизация
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Частица не найдена
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Загрузка изображения частицы
      tags:
      - Particles
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"